Initial upload
This is the beginning of the
first application using Github version control. Only a couple of days of work has gone into it so far.

Writing this application will teach me about utilizing the network in BlitzBASIC.

The server starts and listens for connections, reading anything coming in and dealing with the control code (first 3 letters of the string), and responding depending on what it is.

The client is somewhat a mess. the window is very early in it's design, only there really so I control what it sends to the server. (Clicking "connect to server" is coded, but only works from the second time onwards. closing the window disconnects from the server)